[Feat] CI added
The project structure has been slightly changed and auto-testing has been added
