+++ .github/workflows/main.yml
@@ -1,32 +1,39 @@
-# This is a basic workflow to help you get started with Actions
+name: Continuous Integration
 
-name: CI
-
-# Controls when the workflow will run
 on:
-  # Triggers the workflow on push or pull request events but only for the "main" branch
   push:
     branches: [ "main" ]
-  # Allows you to run this workflow manually from the Actions tab
+  pull_request:
+    branches: [ "main" ]
+
   workflow_dispatch:
 
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
-  # This workflow contains a single job called "build"
-  build:
-    # The type of runner that the job will run on
-    runs-on: [ubuntu-latest, windows-latest]
+  test_ubuntu:
+    runs-on: ubuntu-latest
 
-    # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      - uses: actions/checkout@v3
-      
-      - name: Set up Python
-        uses: actions/setup-python@v2
-        with:
-          python-version: 3.9
+      - uses: actions/checkout@v4
 
-      - name: Run unit tests
+      - uses: actions/setup-python@v4
+        with:
+          python-version: "3.x"
+      
+      - name: Run tests
         run: |
-          python -m pytest tests/
+          pip install pytest
+          pytest
+          
+  test_windows:
+    runs-on: windows-latest
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: actions/setup-python@v4
+        with:
+          python-version: "3.x"
+      
+      - name: Run tests
+        run: |
+          python test_start.py
