+++ test_start.py
@@ -1,0 +1,7 @@
+if __name__ == "__main__":
+    from tests.circle_test import *
+    from tests.rectangle_test import *
+    from tests.square_test import *
+    from tests.triangle_test import *
+
+    unittest.main()
