--- tests/tests_start.py
@@ -1,6 +1,0 @@
-import subprocess
-
-subprocess.run('python rectangle_tests.py', shell=True)
-subprocess.run('python triangle_tests.py', shell=True)
-subprocess.run('python circle_tests.py', shell=True)
-subprocess.run('python square_tests.py', shell=True)
